Favorites updated, temporarly not working
