--- web/src/main/java/com/infoshare/eventmanagers/servlet/FavoriteServlet.java
@@ -1,19 +1,0 @@
-package com.infoshare.eventmanagers.servlet;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@WebServlet("/favorite")
-public class FavoriteServlet extends HttpServlet {
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        int id = Integer.parseInt(request.getParameter("id"));
-        PrintWriter printWriter = response.getWriter();
-        printWriter.println(id);
-    }
-}
