+++ web/src/main/java/com/infoshare/eventmanagers/servlets/FavoriteManageServlet.java
@@ -1,25 +1,15 @@
-package com.infoshare.eventmanagers.servlet;
+package com.infoshare.eventmanagers.servlets;
 
-import com.infoshare.eventmanagers.dto.EventDto;
-import com.infoshare.eventmanagers.providers.TemplateProvider;
-import com.infoshare.eventmanagers.service.FavoriteService;
+import com.infoshare.eventmanagers.services.FavoriteService;
 import com.infoshare.eventmanagers.services.EventService;
-import freemarker.template.Configuration;
-import freemarker.template.Template;
-import freemarker.template.TemplateException;
-import freemarker.template.TemplateExceptionHandler;
 
 import javax.inject.Inject;
-import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.Writer;
-import java.util.HashMap;
-import java.util.Map;
 
 @WebServlet("/favoriteManager")
 public class FavoriteManageServlet extends HttpServlet {
